requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
click==8.3.1
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.122.0
graphql-core==3.2.7
h11==0.16.0
idna==3.11
isort==7.0.0
lia-web==0.2.3
motor==3.7.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
pymongo-migrate==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
pytokens==0.3.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
strawberry-graphql==0.287.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
